+++ crates/load-balancer/src/usage.rs
@@ -185,9 +185,9 @@ fn write_records_to_db(
     conn.execute_batch(
         r#"
         CREATE TABLE IF NOT EXISTS Usage (
-            account_id BIGINT NOT NULL,
-            key_id BIGINT NOT NULL,
-            plan_id BIGINT NOT NULL,
+            account_id INTEGER NOT NULL,
+            key_id INTEGER NOT NULL,
+            plan_id INTEGER NOT NULL,
             date_time DATETIME NOT NULL,
             total_requests INTEGER,
             total_data_mb REAL,
@@ -326,9 +326,9 @@ impl UsageWriter {
         conn.execute_batch(
             r#"
             CREATE TABLE IF NOT EXISTS Usage (
-                account_id BIGINT NOT NULL,
-                key_id BIGINT NOT NULL,
-                plan_id BIGINT NOT NULL,
+                account_id INTEGER NOT NULL,
+                key_id INTEGER NOT NULL,
+                plan_id INTEGER NOT NULL,
                 date_time DATETIME NOT NULL,
                 total_requests INTEGER,
                 total_data_mb REAL,
